requirements.txt
feedparser>=6.0.0
feedgen>=0.9.0
PyYAML>=6.0
requests>=2.28
beautifulsoup4>=4.12
